+++ templates/tutorialv2/messages/marked_obsolete_alert.md
@@ -3,10 +3,8 @@
 {% blocktrans with title=title|safe url=url |safe reason=reason %}
 Bonjour,
 
-Votre contenu « [{{ title }}]({{ url }}) » a été marqué comme obsolète.
-
-**Justification :**
-{{ reason }}
+Votre contenu « [{{ title }}]({{ url }}) » a été marqué comme obsolète avec la précision suivante :
+> {{ reason }}
 
 Si vous en avez la possibilité, pourriez-vous mettre à jour votre contenu ?
 {% endblocktrans %}
